This commit marks the final push before taking a break from the project for at least two weeks. The focus of this push was on improving the Lua script for data extraction. We encountered an issue with pointer arithmetic when trying to access the animation address. This is crucial because without the animation address, we risk collecting excessive NPC data, as Elden Ring doesn't destroy inactive characters but instead sets their animation to a value less than zero.
For now, the player data extraction works, and the project is paused for other endeavors. The data script is hooked up, and most of it functions correctly in Cheat Engine, though there are some tweaks left regarding NPC data extraction. A few changes were made to naming conventions as well.

Notable updates:
- The state machine now stores a reference to the data collector.
- The timer now supports start, stop, and monitoring states, with the ability to close.
- NPC data extraction is in progress.
- Next steps involve hooking up the data to the web request and integrating with the Node.js server.

This is the hardest part of the project, and it’s almost done.